@@ -4,14 +4,25 @@
     <modelVersion>4.0.0</modelVersion>
     <groupId>com.lv.playtube</groupId>
     <artifactId>playtube</artifactId>
-    <version>0.0.1-SNAPSHOT</version>
+    <version>1.0.0-SNAPSHOT</version>
     <name>playtube</name>
     <description>playtube</description>
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>17</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <spring-boot.version>2.6.13</spring-boot.version>
+        <spring-boot.version>3.0.7</spring-boot.version>
+        <mybatis-spring-boot-starter.version>3.0.2</mybatis-spring-boot-starter.version>
+        <rocketmq-client.version>4.9.1</rocketmq-client.version>
+        <rocketmq-spring-boot-starter.version>2.2.2</rocketmq-spring-boot-starter.version>
+        <fastdfs-client.version>1.27.2</fastdfs-client.version>
+        <mysql-connector-java.version>8.0.28</mysql-connector-java.version>
+        <commons-codec.version>1.14</commons-codec.version>
+        <java-jwt.version>3.19.2</java-jwt.version>
+        <fastjson.version>1.2.79</fastjson.version>
+        <junit-jupiter.version>5.9.3</junit-jupiter.version>
+        <transmittable-thread-local.version>2.13.0</transmittable-thread-local.version>
+        <hutool-all.vesion>5.4.3</hutool-all.vesion>
     </properties>
 
     <dependencies>
@@ -130,27 +141,10 @@
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>3.8.1</version>
                 <configuration>
-                    <source>1.8</source>
-                    <target>1.8</target>
+                    <source>17</source>
+                    <target>17</target>
                     <encoding>UTF-8</encoding>
                 </configuration>
-            </plugin>
-            <plugin>
-                <groupId>org.springframework.boot</groupId>
-                <artifactId>spring-boot-maven-plugin</artifactId>
-                <version>${spring-boot.version}</version>
-                <configuration>
-                    <mainClass>com.lv.playtube.PlayTubeApplication</mainClass>
-                    <skip>true</skip>
-                </configuration>
-                <executions>
-                    <execution>
-                        <id>repackage</id>
-                        <goals>
-                            <goal>repackage</goal>
-                        </goals>
-                    </execution>
-                </executions>
             </plugin>
         </plugins>
     </build>
