@@ -2,12 +2,11 @@ package com.playtube.config;
 
 import com.alibaba.fastjson.JSONObject;
 import com.playtube.common.constant.UserMomentsConstant;
+import com.playtube.controller.websocket.WebSocketService;
 import com.playtube.pojo.UserFollowing;
 import com.playtube.pojo.UserMoments;
-import com.playtube.service.impl.UserFollowingServiceImpl;
-import com.playtube.controller.websocket.WebSocketService;
+import com.playtube.service.UserFollowingService;
 import org.apache.rocketmq.client.consumer.DefaultMQPushConsumer;
-import org.apache.rocketmq.client.consumer.listener.ConsumeConcurrentlyContext;
 import org.apache.rocketmq.client.consumer.listener.ConsumeConcurrentlyStatus;
 import org.apache.rocketmq.client.consumer.listener.MessageListenerConcurrently;
 import org.apache.rocketmq.client.producer.DefaultMQProducer;
@@ -24,13 +23,13 @@ import java.util.List;
 @Configuration
 public class RocketMQConfig {
 
-    @Value("${rocketmq.name-server}")
+    @Value("${rocketmq.nameServer}")
     private String nameServerAddr;
 
     @Autowired
     private RedisTemplate<String, String> redisTemplate;
     @Autowired
-    private UserFollowingServiceImpl userFollowingService;
+    private UserFollowingService userFollowingService;
 
     @Bean("momentProducer")
     public DefaultMQProducer momentProducer() throws Exception {
@@ -45,26 +44,23 @@ public class RocketMQConfig {
         DefaultMQPushConsumer consumer = new DefaultMQPushConsumer(UserMomentsConstant.MOMENTS_GROUP);
         consumer.setNamesrvAddr(nameServerAddr);
         consumer.subscribe(UserMomentsConstant.MOMENTS_TOPIC, "*");//订阅主题与二级主题
-        consumer.registerMessageListener(new MessageListenerConcurrently() {
-            @Override
-            public ConsumeConcurrentlyStatus consumeMessage(List<MessageExt> list, ConsumeConcurrentlyContext consumeConcurrentlyContext) {
-                MessageExt msg = list.get(0);
-                if (msg == null) {
-                    return ConsumeConcurrentlyStatus.RECONSUME_LATER;
-                }
-                String bodyStr = new String(msg.getBody());
-                UserMoments userMoments = JSONObject.toJavaObject(JSONObject.parseObject(bodyStr), UserMoments.class);
-                Long userId = userMoments.getUserId();
-                List<UserFollowing> userFans = userFollowingService.getUserFans(userId);
-                for (UserFollowing fan : userFans) {
-                    String key = "subscribed-" + fan.getUserId();//定义redis里的key
-                    redisTemplate.opsForList().leftPush(key,bodyStr);
-                    while(redisTemplate.opsForList().size(key) > 500){
-                        redisTemplate.opsForList().rightPop(key);
-                    }
-                }
-                return ConsumeConcurrentlyStatus.CONSUME_SUCCESS;
+        consumer.registerMessageListener((MessageListenerConcurrently) (list, consumeConcurrentlyContext) -> {
+            MessageExt msg = list.get(0);
+            if (msg == null) {
+                return ConsumeConcurrentlyStatus.RECONSUME_LATER;
             }
+            String bodyStr = new String(msg.getBody());
+            UserMoments userMoments = JSONObject.toJavaObject(JSONObject.parseObject(bodyStr), UserMoments.class);
+            Long userId = userMoments.getUserId();
+            List<UserFollowing> userFans = userFollowingService.getUserFans(userId);
+            for (UserFollowing fan : userFans) {
+                String key = "subscribed-" + fan.getUserId();//定义redis里的key
+                redisTemplate.opsForList().leftPush(key,bodyStr);
+                while(redisTemplate.opsForList().size(key) > 500){
+                    redisTemplate.opsForList().rightPop(key);
+                }
+            }
+            return ConsumeConcurrentlyStatus.CONSUME_SUCCESS;
         });
         consumer.start();
         return consumer;
@@ -86,27 +82,25 @@ public class RocketMQConfig {
         consumer.setNamesrvAddr(nameServerAddr);
         //订阅一个或多个Topic，以及Tag来过滤需要消费的信息
         consumer.subscribe(UserMomentsConstant.BARRAGE_TOPIC, "*");//订阅主题与二级主题
-        consumer.registerMessageListener(new MessageListenerConcurrently() {
-            @Override
-            public ConsumeConcurrentlyStatus consumeMessage(List<MessageExt> list, ConsumeConcurrentlyContext consumeConcurrentlyContext) {
-                MessageExt msg = list.get(0);
-                if (msg == null) {
-                    return ConsumeConcurrentlyStatus.CONSUME_SUCCESS;
-                }
-                String bodyStr = new String(msg.getBody());
-                JSONObject jsonObject = JSONObject.parseObject(bodyStr);
-                String sessionId = jsonObject.getString("sessionId");
-                String message = jsonObject.getString("message");
-                WebSocketService webSocketService = WebSocketService.WEBSOCKET_MAP.get(sessionId);
-                if (webSocketService.getSession().isOpen()) {
-                    try {
-                        webSocketService.sendMessage(message);
-                    } catch (Exception e) {
-                        e.printStackTrace();
-                    }
-                }
+        consumer.registerMessageListener((MessageListenerConcurrently) (list, consumeConcurrentlyContext) -> {
+            MessageExt msg = list.get(0);
+            if (msg == null) {
                 return ConsumeConcurrentlyStatus.CONSUME_SUCCESS;
             }
+            String bodyStr = new String(msg.getBody());
+            JSONObject jsonObject = JSONObject.parseObject(bodyStr);
+            Long videoId = Long.valueOf((String) jsonObject.get("videoId"));
+            String sessionId = (String) jsonObject.get("sessionId");
+            String message = (String) jsonObject.get("message");
+            WebSocketService webSocketService = WebSocketService.WEBSOCKET_MAP.get(videoId).get(sessionId);
+            if (webSocketService.getSession().isOpen()) {
+                try {
+                    webSocketService.sendMessage(message);
+                } catch (Exception e) {
+                    return ConsumeConcurrentlyStatus.RECONSUME_LATER;
+                }
+            }
+            return ConsumeConcurrentlyStatus.CONSUME_SUCCESS;
         });
         consumer.start();
         return consumer;
