@@ -5,9 +5,10 @@ import com.playtube.pojo.Video;
 import com.playtube.pojo.VideoCoin;
 import com.playtube.pojo.VideoCollection;
 import com.playtube.pojo.VideoComment;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+
 import java.util.Map;
 
 public interface VideoService {
