@@ -8,6 +8,7 @@ import com.playtube.common.constant.UserConstant;
 import com.playtube.dao.UserDao;
 import com.playtube.common.exception.ConditionException;
 import com.playtube.pojo.*;
+import com.playtube.service.UserRoleService;
 import com.playtube.service.UserService;
 import com.playtube.util.MD5Util;
 import com.playtube.util.RSAUtil;
@@ -25,7 +26,7 @@ import java.util.concurrent.TimeUnit;
 @RequiredArgsConstructor
 public class UserServiceImpl implements UserService {
     private final UserDao userDao;
-    private final UserRoleServiceImpl userRoleService;
+    private final UserRoleService userRoleService;
     private final RedisTemplate<String,String> redisTemplate;
 
     /**
